requirements.txt
openai>=1.12.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0